gcd, Euler's function ver2'

+++ gcd.cpp
@@ -1,6 +1,6 @@
 ll gcd(ll a, ll b){
 	while(b!=0){
-		int r = a%b;
+		ll r = a%b;
 		a= b;
 		b= r;
 	}
